+++ gdal/ogr/ogrsf_frmts/ntf/ntfrecord.cpp
@@ -159,7 +159,7 @@ int NTFRecord::ReadPhysicalLine( FILE *fp, char *pszLine )
 /*      Read enough data that we are sure we have a whole record.       */
 /* -------------------------------------------------------------------- */
     nRecordStart = VSIFTell( fp );
-    nBytesRead = VSIFRead( pszLine, 1, MAX_RECORD_LEN+2, fp );
+    nBytesRead = static_cast<int>(VSIFRead( pszLine, 1, MAX_RECORD_LEN+2, fp ));
 
     if( nBytesRead == 0 )
     {
